backend/requirements.txt
Flask==2.3.3
supabase==1.0.3
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0